Fix "all" the clippy warnings
