+++ testing/Cargo.toml
@@ -8,7 +8,7 @@ publish = false
 
 [features]
 default = ["serde-json", "markdown", "localization"]
-localization = ["unic-langid", "askama/localization"]
+localization = ["askama/localization"]
 markdown = ["comrak", "askama/markdown"]
 serde-json = ["serde_json", "askama/serde-json"]
 
@@ -16,7 +16,6 @@ serde-json = ["serde_json", "askama/serde-json"]
 askama = { path = "../askama", version = "0.11.0-beta.1" }
 comrak = { version = "0.13", default-features = false, optional = true }
 serde_json = { version = "1.0", optional = true }
-unic-langid = { version =  "0.9.0", optional = true }
 
 [dev-dependencies]
 criterion = "0.3"
